refactor: extract password-field logic into reusable mixin
