+++ packages/password-field/package.json
@@ -34,13 +34,15 @@
     "web-component"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/button": "24.4.0-alpha1",
     "@vaadin/component-base": "24.4.0-alpha1",
     "@vaadin/text-field": "24.4.0-alpha1",
     "@vaadin/vaadin-lumo-styles": "24.4.0-alpha1",
     "@vaadin/vaadin-material-styles": "24.4.0-alpha1",
-    "@vaadin/vaadin-themable-mixin": "24.4.0-alpha1"
+    "@vaadin/vaadin-themable-mixin": "24.4.0-alpha1",
+    "lit": "^3.0.0"
   },
   "devDependencies": {
     "@esm-bundle/chai": "^4.3.4",
