+++ packages/password-field/src/vaadin-password-field-mixin.d.ts
@@ -1,0 +1,37 @@
+/**
+ * @license
+ * Copyright (c) 2021 - 2023 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import type { Constructor } from '@open-wc/dedupe-mixin';
+
+export declare function PasswordFieldMixin<T extends Constructor<HTMLElement>>(
+  base: T,
+): Constructor<PasswordFieldMixinClass> & T;
+
+export declare class PasswordFieldMixinClass {
+  /**
+   * Set to true to hide the eye icon which toggles the password visibility.
+   * @attr {boolean} reveal-button-hidden
+   */
+  revealButtonHidden: boolean;
+
+  /**
+   * True if the password is visible ([type=text]).
+   * @attr {boolean} password-visible
+   */
+  readonly passwordVisible: boolean;
+
+  /**
+   * An object with translated strings used for localization.
+   * It has the following structure and default values:
+   *
+   * ```
+   * {
+   *   // Translation of the reveal icon button accessible label
+   *   reveal: 'Show password'
+   * }
+   * ```
+   */
+  i18n: { reveal: string };
+}
